Rearrange the files and get the server to work with the routes
