+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 dist
 *.log
 yarn.lock
+src/server/config/default.json
