+++ package.json
@@ -18,6 +18,7 @@
     "axios": "^0.21.0",
     "babel-polyfill": "^6.26.0",
     "bcrypt": "^5.0.0",
+    "config": "^3.3.2",
     "cors": "^2.8.5",
     "express": "^4.16.3",
     "joi": "^17.3.0",
