+++ src/server/config/default.example.json
@@ -1,0 +1,4 @@
+{
+    "mongoURI": "CONNECT URI HERE",
+    "jwtSecret": "JWT SECRET HERE"
+}
